requirements-dev.txt
heatzypy
pytest-homeassistant-custom-component
async-upnp-client